+++ packages/jest-runtime/src/__tests__/runtime_require_mock.test.js
@@ -173,5 +173,14 @@ describe('Runtime', () => {
         expect(exports.isManualMockModule).toBe(true);
       });
     });
+    it('provides `require.main` in mock', () =>
+      createRuntime(__filename).then(runtime => {
+        runtime._moduleRegistry[__filename] = module;
+        runtime.setMock(__filename, 'export_main', () => require.main, {
+          virtual: true,
+        });
+        const mainModule = runtime.requireMock(__filename, 'export_main');
+        expect(mainModule).toBe(module);
+      }));
   });
 });